Moved non-route functions from networkRoutes.js and careerRoutes.js to public/scripts.
