+++ views/careerQuiz.ejs
@@ -64,6 +64,7 @@
 
 </div>
 
+<script src="/scripts/career/careerQuiz.js"></script>
 <script>
     const steps = document.querySelectorAll('.question-step');
     let currentStep = 0;
