+++ views/careers.ejs
@@ -1,52 +1,60 @@
 <!DOCTYPE html>
 <html lang="en">
-<!--<head>-->
-<!--    <meta charset="UTF-8">-->
-<!--    <meta name="viewport" content="width=device-width, initial-scale=1.0">-->
-<!--    <title>Career Planner</title>-->
-<!--    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">-->
-<!--    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.0/font/bootstrap-icons.css">-->
-<!--    <link rel="stylesheet" href="">-->
-<!--</head>-->
-<%- include('templates/header',{css:["css/index.css","/css/careers.css"]}) %>
-<body>
+  <!--<head>-->
+  <!--    <meta charset="UTF-8">-->
+  <!--    <meta name="viewport" content="width=device-width, initial-scale=1.0">-->
+  <!--    <title>Career Planner</title>-->
+  <!--    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">-->
+  <!--    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.0/font/bootstrap-icons.css">-->
+  <!--    <link rel="stylesheet" href="">-->
+  <!--</head>-->
+  <%- include('templates/header',{css:["css/index.css","/css/careers.css"]}) %>
+  <body>
     <div class="header text-center">
-        <div class="container">
-            <h1 class="display-4 fw-bold pa
-            ge-title">Career Planner</h1>
-            <p class="lead">Discover your future in tech</p>
-            <a href="/career_quiz" class="btn btn-explore">Take A Career Quiz</a>
-            <div class="row justify-content-center mt-4">
-                <div class="col-md-8">
-                    <div class="search-box d-flex align-items-center">
-                        <i class="bi bi-search me-2 text-muted"></i>
-                        <input type="text" class="search-input" placeholder="Search careers..." id="searchInput">
-                    </div>
-                </div>
+      <div class="container">
+        <h1 class="display-4 fw-bold pa ge-title">Career Planner</h1>
+        <p class="lead">Discover your future in tech</p>
+        <a href="/career_quiz" class="btn btn-explore">Take A Career Quiz</a>
+        <div class="row justify-content-center mt-4">
+          <div class="col-md-8">
+            <div class="search-box d-flex align-items-center">
+              <i class="bi bi-search me-2 text-muted"></i>
+              <input
+                type="text"
+                class="search-input"
+                placeholder="Search careers..."
+                id="searchInput"
+              />
             </div>
+          </div>
         </div>
+      </div>
     </div>
-    
+
     <div class="container">
-        <div class="row" id="careersContainer">
-            <% careers.forEach(career => { %>
-                <div class="col-lg-4 col-md-6 mb-4 career-item" data-title="<%= career.title.toLowerCase() %>">
-                    <div class="career-card h-100">
-                        <div class="p-4 text-center">
-                            <i class="bi <%= getCareerIcon(career.title) %> card-icon"></i>
-                            <h3 class="card-title"><%= career.title %></h3>
-                            <p class="card-text"><%= career.summary %></p>
-                            <a href="/careers/<%= career.slug %>" class="btn btn-explore">Explore Path</a>
-                        </div>
-                    </div>
-                </div>
-            <% }); %>
+      <div class="row" id="careersContainer">
+        <% careers.forEach(career => { %>
+        <div
+          class="col-lg-4 col-md-6 mb-4 career-item"
+          data-title="<%= career.title.toLowerCase() %>"
+        >
+          <div class="career-card h-100">
+            <div class="p-4 text-center">
+              <i class="bi <%= getCareerIcon(career.title) %> card-icon"></i>
+              <h3 class="card-title"><%= career.title %></h3>
+              <p class="card-text"><%= career.summary %></p>
+              <a href="/careers/<%= career.slug %>" class="btn btn-explore"
+                >Explore Path</a
+              >
+            </div>
+          </div>
         </div>
+        <% }); %>
+      </div>
     </div>
-    
+
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="/scripts/careers.js"></script>
-    
-</body>
-<%- include('templates/footer', {currentPage:'careers'}) %>
+    <script src="/scripts/career/careers.js"></script>
+  </body>
+  <%- include('templates/footer', {currentPage:'careers'}) %>
 </html>
